Rework HttpLink options by moving into namespace
