+++ packages/apollo-angular/http/src/deprecated.ts
@@ -1,0 +1,8 @@
+import { HttpBatchLink } from './http-batch-link';
+import { HttpLink } from './http-link';
+
+/** @deprecated Use `HttpLink.Options` instead */
+export type Options = HttpLink.Options;
+
+/** @deprecated Use `HttpBatchLink.Options` instead */
+export type BatchOptions = HttpBatchLink.Options;
