+++ packages/apollo-angular/http/src/index.ts
@@ -3,4 +3,4 @@ export { HttpLink, HttpLinkHandler } from './http-link';
 // http-batch
 export { HttpBatchLink, HttpBatchLinkHandler } from './http-batch-link';
 // common
-export { BatchOptions, Options } from './types';
+export { BatchOptions, Options } from './deprecated';
